Move ByteBufsInputStream from internal.common to core package
